requirements.txt
redis
cassandra-driver

watchdog
pytest